@@ -24,13 +24,14 @@
         }
     ],
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.2.0",
         "ext-pcre": "*",
         "ext-xml": "*",
         "ext-xmlreader": "*"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.19 || ^3.8",
+        "psr/simple-cache": "^1 || ^2 || ^3",
         "yoast/phpunit-polyfills": "^1.0.1"
     },
     "suggest": {
