@@ -1,7 +1,7 @@
 <?php
 return [
-    'version' => '3.1.13-RC2', // Current version number
-    'release_date' => 'January 1, 2023', // Date of release
+    'version' => '3.1.13', // Current version number
+    'release_date' => 'January 20, 2023', // Date of release
     'branch' => 'Evolution CE', // Codebase name
-    'full_appname' => 'Evolution CE 3.1.13-RC2 (January 1, 2023)'
+    'full_appname' => 'Evolution CE 3.1.13 (January 20, 2023)'
 ];
