@@ -156,7 +156,7 @@ $_lang["configcheck_validate_referer"] = 'Папярэджанне бяспек�
 $_lang["configcheck_validate_referer_msg"] = 'Параметр канфігурацыі <strong>Праверыць загалоўкі HTTP_REFERER выключана?</strong> Мы рэкамендуем уключыць яго. <a href="index.php?a=17">Перайдзіце да параметраў канфігурацыі</a>';
 $_lang["configcheck_warning"] = 'Папярэджанне канфігурацыі:';
 $_lang["configcheck_what"] = 'Што гэта значыць?';
-$_lang["safe_mode_warning"] = 'Safe mode is enabled. Manager functionality is limited.';
+$_lang["safe_mode_warning"] = 'Уключаны бяспечны рэжым. Функцыянальнасць мэнэджара абмежаваная.';
 $_lang["confirm_block"] = 'Вы ўпэўнены, што хочаце заблакіраваць гэтага карыстальніка?';
 $_lang["confirm_delete_category"] = 'Вы ўпэўнены, што хочаце выдаліць гэту катэгорыю?';
 $_lang["confirm_delete_eventlog"] = 'Вы ўпэўнены, што хочаце выдаліць гэты журнал падзей?';
@@ -1104,13 +1104,13 @@ $_lang["update_params"] = 'Адлюстраванне параметраў аб�
 $_lang["update_settings_from_language"] = 'Замяніць бягучы на:';
 $_lang["upload_maxsize_message"] = 'Увядзіце максімальны памер файла, які можна загрузіць праз файлавы менеджэр. Памер загружанага файла павінен быць уведзены ў байтах. <b>УВАГА: загрузка вялікіх файлаў можа заняць вельмі шмат часу!</b>';
 $_lang["upload_maxsize_title"] = 'Максімальны памер загрузкі';
-$_lang["uploadable_files_message"] = 'Enter a list of files that can be uploaded into \'assets/files/\' using the File Browser. Please enter the extensions for the filetypes, separated by commas.';
+$_lang["uploadable_files_message"] = 'Увядзіце спіс файлаў, якія можна загружаць у \'assets/files/\' з дапамогай файлавага браўзера. Калі ласка, увядзіце пашырэнні для тыпаў файлаў праз коскі.';
 $_lang["uploadable_files_title"] = 'Тыпы файлаў, якія можна загрузіць';
-$_lang["uploadable_flash_message"] = 'Enter a list of files that can be uploaded into \'assets/flash/\' using the File Browser. Please enter the extensions for the flash types, separated by commas.';
+$_lang["uploadable_flash_message"] = 'Увядзіце спіс файлаў, якія можна загружаць у \'assets/flash/\' з дапамогай файлавага браўзера. Калі ласка, увядзіце пашырэнні для тыпаў флэш праз коскі.';
 $_lang["uploadable_flash_title"] = 'Загружаныя тыпы Flash';
-$_lang["uploadable_images_message"] = 'Enter a list of files that can be uploaded into \'assets/images/\' using the File Browser. Please enter the extensions for the image types, separated by commas.';
+$_lang["uploadable_images_message"] = 'Увядзіце спіс файлаў, якія можна загружаць у \'assets/images/\' з дапамогай файлавага браўзера. Калі ласка, увядзіце пашырэнні для тыпаў малюнкаў праз коскі.';
 $_lang["uploadable_images_title"] = 'Тыпы малюнкаў, якія можна загрузіць';
-$_lang["uploadable_media_message"] = 'Enter a list of files that can be uploaded into \'assets/media/\' using the File Browser. Please enter the extensions for the media types, separated by commas.';
+$_lang["uploadable_media_message"] = 'Увядзіце спіс файлаў, якія можна загружаць у \'assets/media/\' з дапамогай файлавага браўзера. Калі ласка, увядзіце пашырэнні для тыпаў медыя-файлаў праз коскі.';
 $_lang["uploadable_media_title"] = 'Тыпы носьбітаў, якія можна загружаць';
 $_lang["use_alias_path_message"] = 'Уключэнне гэтай опцыі стварае віртуальны шлях да рэсурсу. Напрыклад, калі рэсурс "child.html" знаходзіцца ў кантэйнерным рэсурсе "parent", то поўны URL-адрас псеўданіма шляху будзе "/parent/child.html".';
 $_lang["use_alias_path_title"] = 'Выкарыстоўвайце дружалюбны URL-псеўданім';
@@ -1168,7 +1168,7 @@ $_lang["view_log"] = 'Прагляд часопіса';
 $_lang["view_logging"] = 'Дзеянні кіраўніка';
 $_lang["view_sysinfo"] = 'Сістэмная інфармацыя';
 $_lang["warning"] = 'Папярэджанне!';
-$_lang["warning_not_saved"] = 'The changes you have made have not been saved yet. You can choose to stay on the current page in order to save the changes (\'Cancel\'), or you can leave this page, losing any changes you have made (\'OK\').';
+$_lang["warning_not_saved"] = 'Унесеныя вамі змены яшчэ не захаваны. Вы можаце застацца на бягучай старонцы, каб захаваць змены (\'Адмена\'), або вы можаце пакінуць гэтую старонку, страціўшы ўсе зробленыя вамі змены (\'ОК\').';
 $_lang["warning_visibility"] = 'Папярэджанні канфігурацыі бачныя';
 $_lang["warning_visibility_message"] = 'Кантралюйце бачнасць папярэджанняў канфігурацыі, якія паказваюцца на старонцы прывітання Менеджара';
 $_lang["web_access_permissions"] = 'Правы доступу';
@@ -1296,11 +1296,11 @@ $_lang["cache_type_2"] = 'Кэш заснаваны на параметрах Re
 $_lang["seostrict_title"] = 'Выкарыстоўвайце строгія URL-адрасы SEO';
 $_lang["seostrict_message"] = 'Захоўвае выкарыстанне строгіх URL-адрасоў для прадухілення дублявання кантэнту (пры неабходнасці)';
 
-$_lang["alias_listing_title"] = 'Use AliasListing cache';
-$_lang["alias_listing_message"] = 'Caching page aliases, have to be disabled if a site have huge amount of resources. "Disabled" reduces memory consumption when site have large number of resources.';
-$_lang["alias_listing_disabled"] = 'Disabled';
-$_lang["alias_listing_folders"] = 'Only for folders';
-$_lang["alias_listing_enabled"] = 'Enabled';
+$_lang["alias_listing_title"] = 'Выкарыстоўваць кэш AliasListing';
+$_lang["alias_listing_message"] = 'Кэшаванне псеўданімаў старонак трэба адключыць, калі сайт мае вялікую колькасць рэсурсаў. "Адключана" памяншае спажыванне памяці, калі сайт мае вялікую колькасць рэсурсаў.';
+$_lang["alias_listing_disabled"] = 'Адключана';
+$_lang["alias_listing_folders"] = 'Толькі для папак';
+$_lang["alias_listing_enabled"] = 'Уключана';
 
 $_lang["settings_friendlyurls_alert"] = 'Неабходна перайменаваць файл ht.access у каталогу ўстаноўкі Evolution CMS у .htaccess, каб выкарыстоўваць функцыю Friendly URL.';
 $_lang["settings_friendlyurls_alert2"] = 'Паколькі Evolution CMS была ўсталявана ў падкаталог, неабходна змяніць змесціва .htaccess.';
