@@ -139,7 +139,7 @@ $_lang["configcheck_notok"] = 'Конфигурация содержит оши�
 $_lang["configcheck_ok"] = 'Конфигурация не содержит ошибок.';
 $_lang["configcheck_php_gdzip"] = 'GD и/или Zip PHP расширения не найдены';
 $_lang["configcheck_php_gdzip_msg"] = 'Для нормальной работы Evolution CMS необходимо, чтобы были разрешены GD и Zip расширения для PHP. Evolution CMS будет работать и без этих расширений, но вы не сможете использовать все возможности встроенного файл-менеджера, редактора изображений или код CAPTCHA для авторизации.';
-$_lang["configcheck_rb_base_dir"] = 'Указан неверный <a href="/[+MGR_DIR+]/?a=17&amp;tab=5">путь для файл-браузера</a>';
+$_lang["configcheck_rb_base_dir"] = 'Указан неверный <a href="/[+MGR_DIR+]/?a=17&amp;tab=4">путь для файл-браузера</a>';
 $_lang["configcheck_rb_base_dir_msg"] = 'Это может произойти, например, после перемещения вашего сайта в другой каталог или на другой сервер. Пожалуйста, проверьте и обновите настройки вашей Evolution CMS';
 $_lang["configcheck_register_globals"] = 'Параметр \'register_globals\' имеет значение \'ON\' в конфигурационном файле \'php.ini\'.';
 $_lang["configcheck_register_globals_msg"] = 'Такая конфигурация делает ваш сайт значительно более уязвимым для Cross Site Scripting (XSS) атак. Обратитесь в службу поддержки вашего хостинга за информацией, каким образом выключить этот параметр.';
