@@ -67,9 +67,9 @@ class AuthServices
     /**
      * Get User
      *
-     * @return User
+     * @return User|null
      */
-    public function user(): User
+    public function user(): ?User
     {
         return $this->user;
     }
