@@ -14,20 +14,20 @@
 $_lang["No errors occured."] = "No se produjeron los errores.";
 $_lang["An error occured!"] = "¡Ocurrió un error!";
 $_lang["You don't have enough privileges for this action!"] = "¡No tiene usted suficientes privilegios para esta acción!";
-$_lang["ID passed in request is NaN!"] = "¡El ID pasado en la solicitud es NaN!";
-$_lang["The document is locked!"] = "¡El recurso está bloqueado!";
+$_lang["ID passed in request is NaN!"] = "¡ID pasado en la solicitud es NaN!";
+$_lang["The document is locked!"] = "¡Recurso está bloqueado!";
 $_lang["Too many results returned from database!"] = "¡Demasiados resultados devueltos de la base de datos!";
 $_lang["Not enough/ no results returned from database!"] = "¡No es suficiente/no se devuelven resultados de la base de datos!";
 $_lang["Couldn't find parent document's name!"] = "¡No se pudo encontrar el nombre de recurso principal!";
 $_lang["Logging error!"] = "¡Error de registro!";
 $_lang["Table to optimise not found in request!"] = "¡No se encontró la tabla para optimizar en la solicitud!";
 $_lang["No settings found in request!"] = "¡No se encontraron configuraciones en la solicitud!";
-$_lang["The document must have a title!"] = "¡El documento debe tener un título!";
+$_lang["The document must have a title!"] = "¡Recurso debe tener un título!";
 $_lang["No user selected as recipient of this message!"] = "¡Ningún usuario está seleccionado como el destinatario de este mensaje!";
 $_lang["No group selected as recipient of this message!"] = "¡Ningún grupo está seleccionado como el destinatario de este mensaje!";
 $_lang["The document was not found!"] = "¡No se encontró el recurso!";
 $_lang["Double action (GET & POST) posted!"] = "¡Doble acción (GET & POST) está publicada!";
-$_lang["Document cannot be it's own parent!"] = "¡El recurso no puede ser su propio padre!";
+$_lang["Document cannot be it's own parent!"] = "¡Recurso no puede ser su propio padre!";
 $_lang["Document's ID not passed in request!"] = "¡No se ha pasado el ID de recurso en la solicitud!";
 $_lang["New parent not set in request!"] = "El principal nuevo no está establecido en la solicitud!";
 $_lang["don't know the user!" ] = "¡Se ha introducido un nombre de usuario o una contraseña incorrectos o ha sido bloqueado!";
