@@ -19,16 +19,16 @@ $_lang["The document is locked!"] = "Документ заблокований!"
 $_lang["Too many results returned from database!"] = "Забагато результатів повернуто з бази даних!";
 $_lang["Not enough/ no results returned from database!"] = "Недостатньо/немає результатів з бази даних!";
 $_lang["Couldn't find parent document's name!"] = "Не вдалося знайти назву батьківського документа!";
-$_lang["Logging error!"] = "Logging error!";
-$_lang["Table to optimise not found in request!"] = "Table to optimise not found in request!";
-$_lang["No settings found in request!"] = "No settings found in request!";
-$_lang["The document must have a title!"] = "The document must have a title!";
-$_lang["No user selected as recipient of this message!"] = "No user selected as recipient of this message!";
-$_lang["No group selected as recipient of this message!"] = "No group selected as recipient of this message!";
-$_lang["The document was not found!"] = "The document was not found!";
-$_lang["Double action (GET & POST) posted!"] = "Double action (GET & POST) posted!";
-$_lang["Document cannot be it's own parent!"] = "Document cannot be it's own parent!";
-$_lang["Document's ID not passed in request!"] = "Document's ID not passed in request!";
+$_lang["Logging error!"] = "Помилка логування!";
+$_lang["Table to optimise not found in request!"] = "Таблиця для оптимізації не знайдена в запиті!";
+$_lang["No settings found in request!"] = "У запиті не знайдено налаштувань!";
+$_lang["The document must have a title!"] = "Документ повинен мати заголовок!";
+$_lang["No user selected as recipient of this message!"] = "Не вибрано жодного користувача як одержувача цього повідомлення!";
+$_lang["No group selected as recipient of this message!"] = "Жодна група не вибрана як одержувач цього повідомлення!";
+$_lang["The document was not found!"] = "Документ не знайдено!";
+$_lang["Double action (GET & POST) posted!"] = "Опубліковано подвійну дію (GET & POST)!";
+$_lang["Document cannot be it's own parent!"] = "Документ не може бути своїм батьком!";
+$_lang["Document's ID not passed in request!"] = "Ідентифікатор документа ID не передано в запит!";
 $_lang["New parent not set in request!"] = "Новий батьківський документ не існує в запиті!";
 $_lang["don't know the user!" ] = "Введено неправильне ім'я користувача або пароль, або вас заблокували!";
 $_lang["wrong password!"] = "Введено неправильне ім'я користувача або пароль, або вас заблокували!";
